+++ flutter_local_notifications/macos/Classes/FlutterLocalNotificationsPlugin.swift
@@ -38,6 +38,7 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
         static let platformSpecifics = "platformSpecifics"
         static let badgeNumber = "badgeNumber"
         static let repeatInterval = "repeatInterval"
+        static let repeatIntervalMilliseconds = "repeatIntervalMilliseconds"
         static let attachments = "attachments"
         static let identifier = "identifier"
         static let filePath = "filePath"
@@ -216,6 +217,8 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
         case "zonedSchedule":
             zonedSchedule(call, result)
         case "periodicallyShow":
+            periodicallyShow(call, result)
+        case "periodicallyShowWithDuration":
             periodicallyShow(call, result)
         default:
             result(FlutterMethodNotImplemented)
@@ -491,21 +494,29 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
         } else {
             let arguments = call.arguments as! [String: AnyObject]
             let notification = buildNSUserNotification(fromArguments: arguments)
-            let rawRepeatInterval = arguments[MethodCallArguments.repeatInterval] as! Int
-            let repeatInterval = RepeatInterval.init(rawValue: rawRepeatInterval)!
-            switch repeatInterval {
-            case .everyMinute:
-                notification.deliveryDate = Date.init(timeIntervalSinceNow: 60)
-                notification.deliveryRepeatInterval = DateComponents.init(minute: 1)
-            case .hourly:
-                notification.deliveryDate = Date.init(timeIntervalSinceNow: 60 * 60)
-                notification.deliveryRepeatInterval = DateComponents.init(hour: 1)
-            case .daily:
-                notification.deliveryDate = Date.init(timeIntervalSinceNow: 60 * 60 * 24)
-                notification.deliveryRepeatInterval = DateComponents.init(day: 1)
-            case .weekly:
-                notification.deliveryDate = Date.init(timeIntervalSinceNow: 60 * 60 * 24 * 7)
-                notification.deliveryRepeatInterval = DateComponents.init(weekOfYear: 1)
+            let rawRepeatInterval = arguments[MethodCallArguments.repeatInterval] as? Int
+            let repeatIntervalMilliseconds = arguments[MethodCallArguments.repeatIntervalMilliseconds] as? Int
+            
+            if(rawRepeatInterval != nil) {
+                let repeatInterval = RepeatInterval.init(rawValue: rawRepeatInterval!)!
+                switch repeatInterval {
+                case .everyMinute:
+                    notification.deliveryDate = Date.init(timeIntervalSinceNow: 60)
+                    notification.deliveryRepeatInterval = DateComponents.init(minute: 1)
+                case .hourly:
+                    notification.deliveryDate = Date.init(timeIntervalSinceNow: 60 * 60)
+                    notification.deliveryRepeatInterval = DateComponents.init(hour: 1)
+                case .daily:
+                    notification.deliveryDate = Date.init(timeIntervalSinceNow: 60 * 60 * 24)
+                    notification.deliveryRepeatInterval = DateComponents.init(day: 1)
+                case .weekly:
+                    notification.deliveryDate = Date.init(timeIntervalSinceNow: 60 * 60 * 24 * 7)
+                    notification.deliveryRepeatInterval = DateComponents.init(weekOfYear: 1)
+                }
+            } else if (repeatIntervalMilliseconds != nil) {
+                let repeatIntervalSeconds = repeatIntervalMilliseconds! / 1000
+                notification.deliveryDate = Date.init(timeIntervalSinceNow: TimeInterval(repeatIntervalSeconds))
+                notification.deliveryRepeatInterval = DateComponents.init(second:repeatIntervalSeconds)
             }
             NSUserNotificationCenter.default.scheduleNotification(notification)
             result(nil)
@@ -632,19 +643,25 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
 
     @available(macOS 10.14, *)
     func buildUserNotificationTimeIntervalTrigger(fromArguments arguments: [String: AnyObject]) -> UNTimeIntervalNotificationTrigger {
-        let rawRepeatInterval = arguments[MethodCallArguments.repeatInterval] as! Int
-        let repeatInterval = RepeatInterval.init(rawValue: rawRepeatInterval)!
-        switch repeatInterval {
-        case .everyMinute:
-            return UNTimeIntervalNotificationTrigger.init(timeInterval: 60, repeats: true)
-        case .hourly:
-            return UNTimeIntervalNotificationTrigger.init(timeInterval: 60 * 60, repeats: true)
-        case .daily:
-            return UNTimeIntervalNotificationTrigger.init(timeInterval: 60 * 60 * 24, repeats: true)
-        case .weekly:
-            return UNTimeIntervalNotificationTrigger.init(timeInterval: 60 * 60 * 24 * 7, repeats: true)
+        let rawRepeatInterval = arguments[MethodCallArguments.repeatInterval] as? Int
+        let repeatIntervalMilliseconds = arguments[MethodCallArguments.repeatIntervalMilliseconds] as? Int
+        
+        if(rawRepeatInterval != nil) {
+            let repeatInterval = RepeatInterval.init(rawValue: rawRepeatInterval!)!
+            switch repeatInterval {
+            case .everyMinute:
+                return UNTimeIntervalNotificationTrigger.init(timeInterval: 60, repeats: true)
+            case .hourly:
+                return UNTimeIntervalNotificationTrigger.init(timeInterval: 60 * 60, repeats: true)
+            case .daily:
+                return UNTimeIntervalNotificationTrigger.init(timeInterval: 60 * 60 * 24, repeats: true)
+            case .weekly:
+                return UNTimeIntervalNotificationTrigger.init(timeInterval: 60 * 60 * 24 * 7, repeats: true)
+            }
+        } else {
+            let repeatIntervalSeconds = repeatIntervalMilliseconds! / 1000
+            return UNTimeIntervalNotificationTrigger.init(timeInterval: TimeInterval(repeatIntervalSeconds), repeats: true)
         }
-
     }
 
     @available(macOS 10.14, *)