+++ flutter_local_notifications_platform_interface/CHANGELOG.md
@@ -1,4 +1,8 @@
 
+## [7.2.0]
+
+* Added `periodicallyShowWithDuration()` and corresponding `validateRepeatDurationInterval()` helper method to ensure duration is at least a minute
+
 ## [7.1.0]
 
 * [Android] `bigText` has added to `ActiveNotification` that allows getting information about the longer text associated with a notification displayed using the big text style. Thanks to the PR from [vulpeep](https://github.com/vulpeep)